+++ functions/lib/contracts/types.js
@@ -1,0 +1,6 @@
+"use strict";
+/**
+ * 電子契約モジュール共通の型定義
+ * 設計書: docs/CONTRACTS_DESIGN.md
+ */
+Object.defineProperty(exports, "__esModule", { value: true });